+++ src/compiler/error-handler.ts
@@ -13,14 +13,24 @@ import { KipperSyntaxError } from "../errors";
  * The Error Handler for the Kipper implementation of {@link ANTLRErrorListener}
  */
 export class KipperErrorListener<Token> implements ANTLRErrorListener<Token> {
-  syntaxError(
-    recognizer: Recognizer<Token, any>,
-    offendingSymbol: Token | undefined,
+  syntaxError<T extends Token>(
+    recognizer: Recognizer<T, any>,
+    offendingSymbol: T | undefined,
     line: number,
     charPositionInLine: number,
     msg: string,
     e: RecognitionException | undefined
   ) {
-    throw new KipperSyntaxError<Token>(recognizer, offendingSymbol, line, charPositionInLine, msg, e);
+    // Capitalising the error message
+    msg = msg.charAt(0).toUpperCase() + msg.slice(1);
+
+    throw new KipperSyntaxError<T>(
+      recognizer,
+      offendingSymbol,
+      line,
+      charPositionInLine,
+      msg,
+      e
+    );
   }
 }