+++ src/compiler/file-ctx.ts
@@ -7,6 +7,8 @@
 
 import { CompilationUnitContext } from "./parser";
 import { KipperParseStream } from "./parse-stream";
+import { KipperFileListener } from "./listener";
+import { ParseTreeWalker } from "antlr4ts/tree";
 
 /**
  * The file context class used to represent a file in a compilation.
@@ -30,7 +32,9 @@ export class KipperFileContext {
    */
   private readonly _startItem: CompilationUnitContext;
 
-  constructor(stream: KipperParseStream, startItem: CompilationUnitContext) {
+  constructor(
+    stream: KipperParseStream,
+    startItem: CompilationUnitContext) {
     this._stream = stream;
     this._startItem = startItem;
   }
@@ -38,14 +42,31 @@ export class KipperFileContext {
   /**
    * Returns the {@link KipperParseStream} which contains the raw file data.
    */
-  get stream() {
+  get stream(): KipperParseStream {
     return this._stream;
   }
 
   /**
    * Returns the start item of the parser tree (top item).
    */
-  get startItem() {
+  get startItem(): CompilationUnitContext {
     return this._startItem;
+  }
+
+  /**
+   * Translate the parse tree of this virtual file into an array of valid TypeScript code lines.
+   * @param listener {KipperFileListener} The listener to generate the TypeScript code. The function will return the
+   * generated {@link KipperFileListener.itemBuffer}, which stores the translated code items.
+   */
+  async translate(
+    listener: KipperFileListener = new KipperFileListener(this)
+  ): Promise<Array<string>> {
+    // The walker used to go through the parse tree
+    const walker = new ParseTreeWalker();
+
+    // Walking through the parse tree using the listener
+    walker.walk(listener, this.startItem);
+
+    return listener.itemBuffer;
   }
 }