+++ src/compiler/index.ts
@@ -6,7 +6,7 @@
  */
 export { KipperCompiler, KipperCompileResult } from "./compiler";
 export { KipperErrorListener } from "./error-handler";
-export { KipperStreams, KipperParseStream } from "./parse-stream";
+export { KipperParseStream } from "./parse-stream";
 export { KipperFileListener } from "./listener";
 export { KipperFileContext } from "./file-ctx";
 export * from "./parser";