+++ src/compiler/listener.ts
@@ -4,9 +4,6 @@
  * @copyright 2021-2022 Luna Klatzer
  * @since 0.0.3
  */
-import { KipperListener } from "./parser";
-import { KipperFileContext } from "./file-ctx";
-import { ParserRuleContext } from "antlr4ts";
 import {
   AdditiveExpressionContext,
   ArgumentExpressionListContext,
@@ -14,7 +11,8 @@ import {
   AssignmentOperatorContext,
   BlockItemContext,
   BlockItemListContext,
-  CastOrConvertExpressionContext, CompilationUnitContext,
+  CastOrConvertExpressionContext,
+  CompilationUnitContext,
   CompoundStatementContext,
   ConditionalExpressionContext,
   ConstantExpressionContext,
@@ -25,18 +23,22 @@ import {
   DesignationContext,
   DesignatorContext,
   DesignatorListContext,
-  DirectDeclaratorContext, EndOfItemContext,
+  DirectDeclaratorContext,
+  EndOfItemContext,
   EqualityExpressionContext,
   ExpressionContext,
   ExpressionStatementContext,
   ExternalBlockItemContext,
   ExternalDeclarationContext,
-  ExternalFunctionDefinitionContext, ExternalItemContext, FunctionDefinitionContext,
+  ExternalFunctionDefinitionContext,
+  ExternalItemContext,
+  FunctionDefinitionContext,
   InitDeclaratorContext,
   InitializerContext,
   InitializerListContext,
   IterationStatementContext,
   JumpStatementContext,
+  KipperListener,
   LabeledStatementContext,
   LogicalAndExpressionContext,
   LogicalOrExpressionContext,
@@ -56,6 +58,10 @@ import {
   UnaryExpressionContext,
   UnaryOperatorContext
 } from "./parser";
+import { KipperFileContext } from "./file-ctx";
+import { ParserRuleContext } from "antlr4ts";
+import { TerminalNode } from "antlr4ts/tree/TerminalNode";
+import { ErrorNode } from "antlr4ts/tree/ErrorNode";
 
 /**
  * The listener for a {@link KipperFileContext}, which walks through the generated
@@ -79,16 +85,16 @@ export class KipperFileListener implements KipperListener {
    */
   private readonly _itemBuffer: Array<string>;
 
-  constructor(fileCtx: KipperFileContext, itemBuffer: Array<string>) {
+  constructor(fileCtx: KipperFileContext) {
     this._fileCtx = fileCtx;
-    this._itemBuffer = itemBuffer;
+    this._itemBuffer = [];
   }
 
   /**
    * The {@link KipperFileContext} instance responsible for managing this {@link KipperFileListener} instance
    */
   get fileCtx(): KipperFileContext {
-    return this._fileCtx
+    return this._fileCtx;
   }
 
   /**
@@ -101,620 +107,750 @@ export class KipperFileListener implements KipperListener {
   }
 
   /**
+   * ??? - Let's ignore this for now
+   * @param node The node that was visited
+   */
+  visitTerminal = (/*@NotNull*/ node: TerminalNode): void => {
+  };
+
+  /**
+   * Function that is called when an error node is visited
+   * @param node The node that was visited
+   */
+  visitErrorNode = (/*@NotNull*/ node: ErrorNode): void => {
+  };
+
+  /**
+   * Function that is called every time an time is entered
+   * @param ctx The context of the rule
+   */
+  enterEveryRule = (/*@NotNull*/ ctx: ParserRuleContext): void => {
+  };
+
+  /**
+   * Function that is called every time an item is exited
+   * @param ctx The context of the rule
+   */
+  exitEveryRule = (/*@NotNull*/ ctx: ParserRuleContext): void => {
+  };
+
+  /**
    * Enter a parse tree produced by the `externalFunctionDefinition`
    * labeled alternative in `KipperParser.externalItem`.
    * @param ctx The parse tree (instance of {@link ParserRuleContext})
    */
-  enterExternalFunctionDefinition = (ctx: ExternalFunctionDefinitionContext) => { };
+  enterExternalFunctionDefinition = (ctx: ExternalFunctionDefinitionContext) => {
+  };
 
   /**
    * exit a parse tree produced by the `externalFunctionDefinition`
    * labeled alternative in `KipperParser.externalItem`.
    * @param ctx The parse tree (instance of {@link ParserRuleContext})
    */
-  exitExternalFunctionDefinition = (ctx: ExternalFunctionDefinitionContext) => { };
+  exitExternalFunctionDefinition = (ctx: ExternalFunctionDefinitionContext) => {
+  };
 
   /**
    * Enter a parse tree produced by the `externalDeclaration`
    * labeled alternative in `KipperParser.externalItem`.
    * @param ctx The parse tree (instance of {@link ParserRuleContext})
    */
-  enterExternalDeclaration = (ctx: ExternalDeclarationContext) => { };
+  enterExternalDeclaration = (ctx: ExternalDeclarationContext) => {
+  };
 
   /**
    * exit a parse tree produced by the `externalDeclaration`
    * labeled alternative in `KipperParser.externalItem`.
    * @param ctx The parse tree (instance of {@link ParserRuleContext})
    */
-  exitExternalDeclaration = (ctx: ExternalDeclarationContext) => { };
+  exitExternalDeclaration = (ctx: ExternalDeclarationContext) => {
+  };
 
   /**
    * Enter a parse tree produced by the `externalBlockItem`
    * labeled alternative in `KipperParser.externalItem`.
    * @param ctx The parse tree (instance of {@link ParserRuleContext})
    */
-  enterExternalBlockItem = (ctx: ExternalBlockItemContext) => { };
+  enterExternalBlockItem = (ctx: ExternalBlockItemContext) => {
+  };
 
   /**
    * exit a parse tree produced by the `externalBlockItem`
    * labeled alternative in `KipperParser.externalItem`.
    * @param ctx The parse tree (instance of {@link ParserRuleContext})
    */
-  exitExternalBlockItem = (ctx: ExternalBlockItemContext) => { };
+  exitExternalBlockItem = (ctx: ExternalBlockItemContext) => {
+  };
 
   /**
    * Enter a parse tree produced by `KipperParser.primaryExpression`.
    * @param ctx The parse tree (instance of {@link ParserRuleContext})
    */
-  enterPrimaryExpression = (ctx: PrimaryExpressionContext) => { };
+  enterPrimaryExpression = (ctx: PrimaryExpressionContext) => {
+  };
 
   /**
    * exit a parse tree produced by `KipperParser.primaryExpression`.
    * @param ctx The parse tree (instance of {@link ParserRuleContext})
    */
-  exitPrimaryExpression = (ctx: PrimaryExpressionContext) => { };
+  exitPrimaryExpression = (ctx: PrimaryExpressionContext) => {
+  };
 
   /**
    * Enter a parse tree produced by `KipperParser.postfixExpression`.
    * @param ctx The parse tree (instance of {@link ParserRuleContext})
    */
-  enterPostfixExpression = (ctx: PostfixExpressionContext) => { };
+  enterPostfixExpression = (ctx: PostfixExpressionContext) => {
+  };
 
   /**
    * exit a parse tree produced by `KipperParser.postfixExpression`.
    * @param ctx The parse tree (instance of {@link ParserRuleContext})
    */
-  exitPostfixExpression = (ctx: PostfixExpressionContext) => { };
+  exitPostfixExpression = (ctx: PostfixExpressionContext) => {
+  };
 
   /**
    * Enter a parse tree produced by `KipperParser.argumentExpressionList`.
    * @param ctx The parse tree (instance of {@link ParserRuleContext})
    */
-  enterArgumentExpressionList = (ctx: ArgumentExpressionListContext) => { };
+  enterArgumentExpressionList = (ctx: ArgumentExpressionListContext) => {
+  };
 
   /**
    * exit a parse tree produced by `KipperParser.argumentExpressionList`.
    * @param ctx The parse tree (instance of {@link ParserRuleContext})
    */
-  exitArgumentExpressionList = (ctx: ArgumentExpressionListContext) => { };
+  exitArgumentExpressionList = (ctx: ArgumentExpressionListContext) => {
+  };
 
   /**
    * Enter a parse tree produced by `KipperParser.unaryExpression`.
    * @param ctx The parse tree (instance of {@link ParserRuleContext})
    */
-  enterUnaryExpression = (ctx: UnaryExpressionContext) => { };
+  enterUnaryExpression = (ctx: UnaryExpressionContext) => {
+  };
 
   /**
    * exit a parse tree produced by `KipperParser.unaryExpression`.
    * @param ctx The parse tree (instance of {@link ParserRuleContext})
    */
-  exitUnaryExpression = (ctx: UnaryExpressionContext) => { };
+  exitUnaryExpression = (ctx: UnaryExpressionContext) => {
+  };
 
   /**
    * Enter a parse tree produced by `KipperParser.unaryOperator`.
    * @param ctx The parse tree (instance of {@link ParserRuleContext})
    */
-  enterUnaryOperator = (ctx: UnaryOperatorContext) => { };
+  enterUnaryOperator = (ctx: UnaryOperatorContext) => {
+  };
 
   /**
    * exit a parse tree produced by `KipperParser.unaryOperator`.
    * @param ctx The parse tree (instance of {@link ParserRuleContext})
    */
-  exitUnaryOperator = (ctx: UnaryOperatorContext) => { };
+  exitUnaryOperator = (ctx: UnaryOperatorContext) => {
+  };
 
   /**
    * Enter a parse tree produced by `KipperParser.castOrConvertExpression`.
    * @param ctx The parse tree (instance of {@link ParserRuleContext})
    */
-  enterCastOrConvertExpression = (ctx: CastOrConvertExpressionContext) => { };
+  enterCastOrConvertExpression = (ctx: CastOrConvertExpressionContext) => {
+  };
 
   /**
    * exit a parse tree produced by `KipperParser.castOrConvertExpression`.
    * @param ctx The parse tree (instance of {@link ParserRuleContext})
    */
-  exitCastOrConvertExpression = (ctx: CastOrConvertExpressionContext) => { };
+  exitCastOrConvertExpression = (ctx: CastOrConvertExpressionContext) => {
+  };
 
   /**
    * Enter a parse tree produced by `KipperParser.multiplicativeExpression`.
    * @param ctx The parse tree (instance of {@link ParserRuleContext})
    */
-  enterMultiplicativeExpression = (ctx: MultiplicativeExpressionContext) => { };
+  enterMultiplicativeExpression = (ctx: MultiplicativeExpressionContext) => {
+  };
 
   /**
    * exit a parse tree produced by `KipperParser.multiplicativeExpression`.
    * @param ctx The parse tree (instance of {@link ParserRuleContext})
    */
-  exitMultiplicativeExpression = (ctx: MultiplicativeExpressionContext) => { };
+  exitMultiplicativeExpression = (ctx: MultiplicativeExpressionContext) => {
+  };
 
   /**
    * Enter a parse tree produced by `KipperParser.additiveExpression`.
    * @param ctx The parse tree (instance of {@link ParserRuleContext})
    */
-  enterAdditiveExpression = (ctx: AdditiveExpressionContext) => { };
+  enterAdditiveExpression = (ctx: AdditiveExpressionContext) => {
+  };
 
   /**
    * exit a parse tree produced by `KipperParser.additiveExpression`.
    * @param ctx The parse tree (instance of {@link ParserRuleContext})
    */
-  exitAdditiveExpression = (ctx: AdditiveExpressionContext) => { };
+  exitAdditiveExpression = (ctx: AdditiveExpressionContext) => {
+  };
 
   /**
    * Enter a parse tree produced by `KipperParser.relationalExpression`.
    * @param ctx The parse tree (instance of {@link ParserRuleContext})
    */
-  enterRelationalExpression = (ctx: RelationalExpressionContext) => { };
+  enterRelationalExpression = (ctx: RelationalExpressionContext) => {
+  };
 
   /**
    * exit a parse tree produced by `KipperParser.relationalExpression`.
    * @param ctx The parse tree (instance of {@link ParserRuleContext})
    */
-  exitRelationalExpression = (ctx: RelationalExpressionContext) => { };
+  exitRelationalExpression = (ctx: RelationalExpressionContext) => {
+  };
 
   /**
    * Enter a parse tree produced by `KipperParser.equalityExpression`.
    * @param ctx The parse tree (instance of {@link ParserRuleContext})
    */
-  enterEqualityExpression = (ctx: EqualityExpressionContext) => { };
+  enterEqualityExpression = (ctx: EqualityExpressionContext) => {
+  };
 
   /**
    * exit a parse tree produced by `KipperParser.equalityExpression`.
    * @param ctx The parse tree (instance of {@link ParserRuleContext})
    */
-  exitEqualityExpression = (ctx: EqualityExpressionContext) => { };
+  exitEqualityExpression = (ctx: EqualityExpressionContext) => {
+  };
 
   /**
    * Enter a parse tree produced by `KipperParser.logicalAndExpression`.
    * @param ctx The parse tree (instance of {@link ParserRuleContext})
    */
-  enterLogicalAndExpression = (ctx: LogicalAndExpressionContext) => { };
+  enterLogicalAndExpression = (ctx: LogicalAndExpressionContext) => {
+  };
 
   /**
    * exit a parse tree produced by `KipperParser.logicalAndExpression`.
    * @param ctx The parse tree (instance of {@link ParserRuleContext})
    */
-  exitLogicalAndExpression = (ctx: LogicalAndExpressionContext) => { };
+  exitLogicalAndExpression = (ctx: LogicalAndExpressionContext) => {
+  };
 
   /**
    * Enter a parse tree produced by `KipperParser.logicalOrExpression`.
    * @param ctx The parse tree (instance of {@link ParserRuleContext})
    */
-  enterLogicalOrExpression = (ctx: LogicalOrExpressionContext) => { };
+  enterLogicalOrExpression = (ctx: LogicalOrExpressionContext) => {
+  };
 
   /**
    * exit a parse tree produced by `KipperParser.logicalOrExpression`.
    * @param ctx The parse tree (instance of {@link ParserRuleContext})
    */
-  exitLogicalOrExpression = (ctx: LogicalOrExpressionContext) => { };
+  exitLogicalOrExpression = (ctx: LogicalOrExpressionContext) => {
+  };
 
   /**
    * Enter a parse tree produced by `KipperParser.conditionalExpression`.
    * @param ctx The parse tree (instance of {@link ParserRuleContext})
    */
-  enterConditionalExpression = (ctx: ConditionalExpressionContext) => { };
+  enterConditionalExpression = (ctx: ConditionalExpressionContext) => {
+  };
 
   /**
    * exit a parse tree produced by `KipperParser.conditionalExpression`.
    * @param ctx The parse tree (instance of {@link ParserRuleContext})
    */
-  exitConditionalExpression = (ctx: ConditionalExpressionContext) => { };
+  exitConditionalExpression = (ctx: ConditionalExpressionContext) => {
+  };
 
   /**
    * Enter a parse tree produced by `KipperParser.assignmentExpression`.
    * @param ctx The parse tree (instance of {@link ParserRuleContext})
    */
-  enterAssignmentExpression = (ctx: AssignmentExpressionContext) => { };
+  enterAssignmentExpression = (ctx: AssignmentExpressionContext) => {
+  };
 
   /**
    * exit a parse tree produced by `KipperParser.assignmentExpression`.
    * @param ctx The parse tree (instance of {@link ParserRuleContext})
    */
-  exitAssignmentExpression = (ctx: AssignmentExpressionContext) => { };
+  exitAssignmentExpression = (ctx: AssignmentExpressionContext) => {
+  };
 
   /**
    * Enter a parse tree produced by `KipperParser.assignmentOperator`.
    * @param ctx The parse tree (instance of {@link ParserRuleContext})
    */
-  enterAssignmentOperator = (ctx: AssignmentOperatorContext) => { };
+  enterAssignmentOperator = (ctx: AssignmentOperatorContext) => {
+  };
 
   /**
    * exit a parse tree produced by `KipperParser.assignmentOperator`.
    * @param ctx The parse tree (instance of {@link ParserRuleContext})
    */
-  exitAssignmentOperator = (ctx: AssignmentOperatorContext) => { };
+  exitAssignmentOperator = (ctx: AssignmentOperatorContext) => {
+  };
 
   /**
    * Enter a parse tree produced by `KipperParser.expression`.
    * @param ctx The parse tree (instance of {@link ParserRuleContext})
    */
-  enterExpression = (ctx: ExpressionContext) => { };
+  enterExpression = (ctx: ExpressionContext) => {
+  };
 
   /**
    * exit a parse tree produced by `KipperParser.expression`.
    * @param ctx The parse tree (instance of {@link ParserRuleContext})
    */
-  exitExpression = (ctx: ExpressionContext) => { };
+  exitExpression = (ctx: ExpressionContext) => {
+  };
 
   /**
    * Enter a parse tree produced by `KipperParser.constantExpression`.
    * @param ctx The parse tree (instance of {@link ParserRuleContext})
    */
-  enterConstantExpression = (ctx: ConstantExpressionContext) => { };
+  enterConstantExpression = (ctx: ConstantExpressionContext) => {
+  };
 
   /**
    * exit a parse tree produced by `KipperParser.constantExpression`.
    * @param ctx The parse tree (instance of {@link ParserRuleContext})
    */
-  exitConstantExpression = (ctx: ConstantExpressionContext) => { };
+  exitConstantExpression = (ctx: ConstantExpressionContext) => {
+  };
 
   /**
    * Enter a parse tree produced by `KipperParser.declaration`.
    * @param ctx The parse tree (instance of {@link ParserRuleContext})
    */
-  enterDeclaration = (ctx: DeclarationContext) => { };
+  enterDeclaration = (ctx: DeclarationContext) => {
+  };
 
   /**
    * exit a parse tree produced by `KipperParser.declaration`.
    * @param ctx The parse tree (instance of {@link ParserRuleContext})
    */
-  exitDeclaration = (ctx: DeclarationContext) => { };
+  exitDeclaration = (ctx: DeclarationContext) => {
+  };
 
   /**
    * Enter a parse tree produced by `KipperParser.storageTypeSpecifier`.
    * @param ctx The parse tree (instance of {@link ParserRuleContext})
    */
-  enterStorageTypeSpecifier = (ctx: StorageTypeSpecifierContext) => { };
+  enterStorageTypeSpecifier = (ctx: StorageTypeSpecifierContext) => {
+  };
 
   /**
    * exit a parse tree produced by `KipperParser.storageTypeSpecifier`.
    * @param ctx The parse tree (instance of {@link ParserRuleContext})
    */
-  exitStorageTypeSpecifier = (ctx: StorageTypeSpecifierContext) => { };
+  exitStorageTypeSpecifier = (ctx: StorageTypeSpecifierContext) => {
+  };
 
   /**
    * Enter a parse tree produced by `KipperParser.declarationSpecifiers`.
    * @param ctx The parse tree (instance of {@link ParserRuleContext})
    */
-  enterDeclarationSpecifiers = (ctx: DeclarationSpecifiersContext) => { };
+  enterDeclarationSpecifiers = (ctx: DeclarationSpecifiersContext) => {
+  };
 
   /**
    * exit a parse tree produced by `KipperParser.declarationSpecifiers`.
    * @param ctx The parse tree (instance of {@link ParserRuleContext})
    */
-  exitDeclarationSpecifiers = (ctx: DeclarationSpecifiersContext) => { };
+  exitDeclarationSpecifiers = (ctx: DeclarationSpecifiersContext) => {
+  };
 
   /**
    * Enter a parse tree produced by `KipperParser.declarationSpecifier`.
    * @param ctx The parse tree (instance of {@link ParserRuleContext})
    */
-  enterDeclarationSpecifier = (ctx: DeclarationSpecifierContext) => { };
+  enterDeclarationSpecifier = (ctx: DeclarationSpecifierContext) => {
+  };
 
   /**
    * exit a parse tree produced by `KipperParser.declarationSpecifier`.
    * @param ctx The parse tree (instance of {@link ParserRuleContext})
    */
-  exitDeclarationSpecifier = (ctx: DeclarationSpecifierContext) => { };
+  exitDeclarationSpecifier = (ctx: DeclarationSpecifierContext) => {
+  };
 
   /**
    * Enter a parse tree produced by `KipperParser.initDeclarator`.
    * @param ctx The parse tree (instance of {@link ParserRuleContext})
    */
-  enterInitDeclarator = (ctx: InitDeclaratorContext) => { };
+  enterInitDeclarator = (ctx: InitDeclaratorContext) => {
+  };
 
   /**
    * exit a parse tree produced by `KipperParser.initDeclarator`.
    * @param ctx The parse tree (instance of {@link ParserRuleContext})
    */
-  exitInitDeclarator = (ctx: InitDeclaratorContext) => { };
+  exitInitDeclarator = (ctx: InitDeclaratorContext) => {
+  };
 
   /**
    * Enter a parse tree produced by `KipperParser.typeSpecifier`.
    * @param ctx The parse tree (instance of {@link ParserRuleContext})
    */
-  enterTypeSpecifier = (ctx: TypeSpecifierContext) => { };
+  enterTypeSpecifier = (ctx: TypeSpecifierContext) => {
+  };
 
   /**
    * exit a parse tree produced by `KipperParser.typeSpecifier`.
    * @param ctx The parse tree (instance of {@link ParserRuleContext})
    */
-  exitTypeSpecifier = (ctx: TypeSpecifierContext) => { };
+  exitTypeSpecifier = (ctx: TypeSpecifierContext) => {
+  };
 
   /**
    * Enter a parse tree produced by `KipperParser.declarator`.
    * @param ctx The parse tree (instance of {@link ParserRuleContext})
    */
-  enterDeclarator = (ctx: DeclaratorContext) => { };
+  enterDeclarator = (ctx: DeclaratorContext) => {
+  };
 
   /**
    * exit a parse tree produced by `KipperParser.declarator`.
    * @param ctx The parse tree (instance of {@link ParserRuleContext})
    */
-  exitDeclarator = (ctx: DeclaratorContext) => { };
+  exitDeclarator = (ctx: DeclaratorContext) => {
+  };
 
   /**
    * Enter a parse tree produced by `KipperParser.directDeclarator`.
    * @param ctx The parse tree (instance of {@link ParserRuleContext})
    */
-  enterDirectDeclarator = (ctx: DirectDeclaratorContext) => { };
+  enterDirectDeclarator = (ctx: DirectDeclaratorContext) => {
+  };
 
   /**
    * exit a parse tree produced by `KipperParser.directDeclarator`.
    * @param ctx The parse tree (instance of {@link ParserRuleContext})
    */
-  exitDirectDeclarator = (ctx: DirectDeclaratorContext) => { };
+  exitDirectDeclarator = (ctx: DirectDeclaratorContext) => {
+  };
 
   /**
    * Enter a parse tree produced by `KipperParser.nestedParenthesesBlock`.
    * @param ctx The parse tree (instance of {@link ParserRuleContext})
    */
-  enterNestedParenthesesBlock = (ctx: NestedParenthesesBlockContext) => { };
+  enterNestedParenthesesBlock = (ctx: NestedParenthesesBlockContext) => {
+  };
 
   /**
    * exit a parse tree produced by `KipperParser.nestedParenthesesBlock`.
    * @param ctx The parse tree (instance of {@link ParserRuleContext})
    */
-  exitNestedParenthesesBlock = (ctx: NestedParenthesesBlockContext) => { };
+  exitNestedParenthesesBlock = (ctx: NestedParenthesesBlockContext) => {
+  };
 
   /**
    * Enter a parse tree produced by `KipperParser.parameterTypeList`.
    * @param ctx The parse tree (instance of {@link ParserRuleContext})
    */
-  enterParameterTypeList = (ctx: ParameterTypeListContext) => { };
+  enterParameterTypeList = (ctx: ParameterTypeListContext) => {
+  };
 
   /**
    * exit a parse tree produced by `KipperParser.parameterTypeList`.
    * @param ctx The parse tree (instance of {@link ParserRuleContext})
    */
-  exitParameterTypeList = (ctx: ParameterTypeListContext) => { };
+  exitParameterTypeList = (ctx: ParameterTypeListContext) => {
+  };
 
   /**
    * Enter a parse tree produced by `KipperParser.parameterList`.
    * @param ctx The parse tree (instance of {@link ParserRuleContext})
    */
-  enterParameterList = (ctx: ParameterListContext) => { };
+  enterParameterList = (ctx: ParameterListContext) => {
+  };
 
   /**
    * exit a parse tree produced by `KipperParser.parameterList`.
    * @param ctx The parse tree (instance of {@link ParserRuleContext})
    */
-  exitParameterList = (ctx: ParameterListContext) => { };
+  exitParameterList = (ctx: ParameterListContext) => {
+  };
 
   /**
    * Enter a parse tree produced by `KipperParser.parameterDeclaration`.
    * @param ctx The parse tree (instance of {@link ParserRuleContext})
    */
-  enterParameterDeclaration = (ctx: ParameterDeclarationContext) => { };
+  enterParameterDeclaration = (ctx: ParameterDeclarationContext) => {
+  };
 
   /**
    * exit a parse tree produced by `KipperParser.parameterDeclaration`.
    * @param ctx The parse tree (instance of {@link ParserRuleContext})
    */
-  exitParameterDeclaration = (ctx: ParameterDeclarationContext) => { };
+  exitParameterDeclaration = (ctx: ParameterDeclarationContext) => {
+  };
 
   /**
    * Enter a parse tree produced by `KipperParser.initializer`.
    * @param ctx The parse tree (instance of {@link ParserRuleContext})
    */
-  enterInitializer = (ctx: InitializerContext) => { };
+  enterInitializer = (ctx: InitializerContext) => {
+  };
 
   /**
    * exit a parse tree produced by `KipperParser.initializer`.
    * @param ctx The parse tree (instance of {@link ParserRuleContext})
    */
-  exitInitializer = (ctx: InitializerContext) => { };
+  exitInitializer = (ctx: InitializerContext) => {
+  };
 
   /**
    * Enter a parse tree produced by `KipperParser.initializerList`.
    * @param ctx The parse tree (instance of {@link ParserRuleContext})
    */
-  enterInitializerList = (ctx: InitializerListContext) => { };
+  enterInitializerList = (ctx: InitializerListContext) => {
+  };
 
   /**
    * exit a parse tree produced by `KipperParser.initializerList`.
    * @param ctx The parse tree (instance of {@link ParserRuleContext})
    */
-  exitInitializerList = (ctx: InitializerListContext) => { };
+  exitInitializerList = (ctx: InitializerListContext) => {
+  };
 
   /**
    * Enter a parse tree produced by `KipperParser.designation`.
    * @param ctx The parse tree (instance of {@link ParserRuleContext})
    */
-  enterDesignation = (ctx: DesignationContext) => { };
+  enterDesignation = (ctx: DesignationContext) => {
+  };
 
   /**
    * exit a parse tree produced by `KipperParser.designation`.
    * @param ctx The parse tree (instance of {@link ParserRuleContext})
    */
-  exitDesignation = (ctx: DesignationContext) => { };
+  exitDesignation = (ctx: DesignationContext) => {
+  };
 
   /**
    * Enter a parse tree produced by `KipperParser.designatorList`.
    * @param ctx The parse tree (instance of {@link ParserRuleContext})
    */
-  enterDesignatorList = (ctx: DesignatorListContext) => { };
+  enterDesignatorList = (ctx: DesignatorListContext) => {
+  };
 
   /**
    * exit a parse tree produced by `KipperParser.designatorList`.
    * @param ctx The parse tree (instance of {@link ParserRuleContext})
    */
-  exitDesignatorList = (ctx: DesignatorListContext) => { };
+  exitDesignatorList = (ctx: DesignatorListContext) => {
+  };
 
   /**
    * Enter a parse tree produced by `KipperParser.designator`.
    * @param ctx The parse tree (instance of {@link ParserRuleContext})
    */
-  enterDesignator = (ctx: DesignatorContext) => { };
+  enterDesignator = (ctx: DesignatorContext) => {
+  };
 
   /**
    * exit a parse tree produced by `KipperParser.designator`.
    * @param ctx The parse tree (instance of {@link ParserRuleContext})
    */
-  exitDesignator = (ctx: DesignatorContext) => { };
+  exitDesignator = (ctx: DesignatorContext) => {
+  };
 
   /**
    * Enter a parse tree produced by `KipperParser.statement`.
    * @param ctx The parse tree (instance of {@link ParserRuleContext})
    */
-  enterStatement = (ctx: StatementContext) => { };
+  enterStatement = (ctx: StatementContext) => {
+  };
 
   /**
    * exit a parse tree produced by `KipperParser.statement`.
    * @param ctx The parse tree (instance of {@link ParserRuleContext})
    */
-  exitStatement = (ctx: StatementContext) => { };
+  exitStatement = (ctx: StatementContext) => {
+  };
 
   /**
    * Enter a parse tree produced by `KipperParser.labeledStatement`.
    * @param ctx The parse tree (instance of {@link ParserRuleContext})
    */
-  enterLabeledStatement = (ctx: LabeledStatementContext) => { };
+  enterLabeledStatement = (ctx: LabeledStatementContext) => {
+  };
 
   /**
    * exit a parse tree produced by `KipperParser.labeledStatement`.
    * @param ctx The parse tree (instance of {@link ParserRuleContext})
    */
-  exitLabeledStatement = (ctx: LabeledStatementContext) => { };
+  exitLabeledStatement = (ctx: LabeledStatementContext) => {
+  };
 
   /**
    * Enter a parse tree produced by `KipperParser.compoundStatement`.
    * @param ctx The parse tree (instance of {@link ParserRuleContext})
    */
-  enterCompoundStatement = (ctx: CompoundStatementContext) => { };
+  enterCompoundStatement = (ctx: CompoundStatementContext) => {
+  };
 
   /**
    * exit a parse tree produced by `KipperParser.compoundStatement`.
    * @param ctx The parse tree (instance of {@link ParserRuleContext})
    */
-  exitCompoundStatement = (ctx: CompoundStatementContext) => { };
+  exitCompoundStatement = (ctx: CompoundStatementContext) => {
+  };
 
   /**
    * Enter a parse tree produced by `KipperParser.blockItemList`.
    * @param ctx The parse tree (instance of {@link ParserRuleContext})
    */
-  enterBlockItemList = (ctx: BlockItemListContext) => { };
+  enterBlockItemList = (ctx: BlockItemListContext) => {
+  };
 
   /**
    * exit a parse tree produced by `KipperParser.blockItemList`.
    * @param ctx The parse tree (instance of {@link ParserRuleContext})
    */
-  exitBlockItemList = (ctx: BlockItemListContext) => { };
+  exitBlockItemList = (ctx: BlockItemListContext) => {
+  };
 
   /**
    * Enter a parse tree produced by `KipperParser.blockItem`.
    * @param ctx The parse tree (instance of {@link ParserRuleContext})
    */
-  enterBlockItem = (ctx: BlockItemContext) => { };
+  enterBlockItem = (ctx: BlockItemContext) => {
+  };
 
   /**
    * exit a parse tree produced by `KipperParser.blockItem`.
    * @param ctx The parse tree (instance of {@link ParserRuleContext})
    */
-  exitBlockItem = (ctx: BlockItemContext) => { };
+  exitBlockItem = (ctx: BlockItemContext) => {
+  };
 
   /**
    * Enter a parse tree produced by `KipperParser.expressionStatement`.
    * @param ctx The parse tree (instance of {@link ParserRuleContext})
    */
-  enterExpressionStatement = (ctx: ExpressionStatementContext) => { };
+  enterExpressionStatement = (ctx: ExpressionStatementContext) => {
+  };
 
   /**
    * exit a parse tree produced by `KipperParser.expressionStatement`.
    * @param ctx The parse tree (instance of {@link ParserRuleContext})
    */
-  exitExpressionStatement = (ctx: ExpressionStatementContext) => { };
+  exitExpressionStatement = (ctx: ExpressionStatementContext) => {
+  };
 
   /**
    * Enter a parse tree produced by `KipperParser.selectionStatement`.
    * @param ctx The parse tree (instance of {@link ParserRuleContext})
    */
-  enterSelectionStatement = (ctx: SelectionStatementContext) => { };
+  enterSelectionStatement = (ctx: SelectionStatementContext) => {
+  };
 
   /**
    * exit a parse tree produced by `KipperParser.selectionStatement`.
    * @param ctx The parse tree (instance of {@link ParserRuleContext})
    */
-  exitSelectionStatement = (ctx: SelectionStatementContext) => { };
+  exitSelectionStatement = (ctx: SelectionStatementContext) => {
+  };
 
   /**
    * Enter a parse tree produced by `KipperParser.iterationStatement`.
    * @param ctx The parse tree (instance of {@link ParserRuleContext})
    */
-  enterIterationStatement = (ctx: IterationStatementContext) => { };
+  enterIterationStatement = (ctx: IterationStatementContext) => {
+  };
 
   /**
    * exit a parse tree produced by `KipperParser.iterationStatement`.
    * @param ctx The parse tree (instance of {@link ParserRuleContext})
    */
-  exitIterationStatement = (ctx: IterationStatementContext) => { };
+  exitIterationStatement = (ctx: IterationStatementContext) => {
+  };
 
   /**
    * Enter a parse tree produced by `KipperParser.jumpStatement`.
    * @param ctx The parse tree (instance of {@link ParserRuleContext})
    */
-  enterJumpStatement = (ctx: JumpStatementContext) => { };
+  enterJumpStatement = (ctx: JumpStatementContext) => {
+  };
 
   /**
    * exit a parse tree produced by `KipperParser.jumpStatement`.
    * @param ctx The parse tree (instance of {@link ParserRuleContext})
    */
-  exitJumpStatement = (ctx: JumpStatementContext) => { };
+  exitJumpStatement = (ctx: JumpStatementContext) => {
+  };
 
   /**
    * Enter a parse tree produced by `KipperParser.compilationUnit`.
    * @param ctx The parse tree (instance of {@link ParserRuleContext})
    */
-  enterCompilationUnit = (ctx: CompilationUnitContext) => { };
+  enterCompilationUnit = (ctx: CompilationUnitContext) => {
+  };
 
   /**
    * exit a parse tree produced by `KipperParser.compilationUnit`.
    * @param ctx The parse tree (instance of {@link ParserRuleContext})
    */
-  exitCompilationUnit = (ctx: CompilationUnitContext) => { };
+  exitCompilationUnit = (ctx: CompilationUnitContext) => {
+  };
 
   /**
    * Enter a parse tree produced by `KipperParser.translationUnit`.
    * @param ctx The parse tree (instance of {@link ParserRuleContext})
    */
-  enterTranslationUnit = (ctx: TranslationUnitContext) => { };
+  enterTranslationUnit = (ctx: TranslationUnitContext) => {
+  };
 
   /**
    * exit a parse tree produced by `KipperParser.translationUnit`.
    * @param ctx The parse tree (instance of {@link ParserRuleContext})
    */
-  exitTranslationUnit = (ctx: TranslationUnitContext) => { };
+  exitTranslationUnit = (ctx: TranslationUnitContext) => {
+  };
 
   /**
    * Enter a parse tree produced by `KipperParser.externalItem`.
    * @param ctx The parse tree (instance of {@link ParserRuleContext})
    */
-  enterExternalItem = (ctx: ExternalItemContext) => { };
+  enterExternalItem = (ctx: ExternalItemContext) => {
+  };
 
   /**
    * exit a parse tree produced by `KipperParser.externalItem`.
    * @param ctx The parse tree (instance of {@link ParserRuleContext})
    */
-  exitExternalItem = (ctx: ExternalItemContext) => { };
+  exitExternalItem = (ctx: ExternalItemContext) => {
+  };
 
   /**
    * Enter a parse tree produced by `KipperParser.functionDefinition`.
    * @param ctx The parse tree (instance of {@link ParserRuleContext})
    */
-  enterFunctionDefinition = (ctx: FunctionDefinitionContext) => { };
+  enterFunctionDefinition = (ctx: FunctionDefinitionContext) => {
+  };
 
   /**
    * exit a parse tree produced by `KipperParser.functionDefinition`.
    * @param ctx The parse tree (instance of {@link ParserRuleContext})
    */
-  exitFunctionDefinition = (ctx: FunctionDefinitionContext) => { };
+  exitFunctionDefinition = (ctx: FunctionDefinitionContext) => {
+  };
 
   /**
    * Enter a parse tree produced by `KipperParser.endOfItem`.
    * @param ctx The parse tree (instance of {@link ParserRuleContext})
    */
-  enterEndOfItem = (ctx: EndOfItemContext) => { };
+  enterEndOfItem = (ctx: EndOfItemContext) => {
+  };
 
   /**
    * exit a parse tree produced by `KipperParser.endOfItem`.
    * @param ctx The parse tree (instance of {@link ParserRuleContext})
    */
-  exitEndOfItem = (ctx: EndOfItemContext) => { };
+  exitEndOfItem = (ctx: EndOfItemContext) => {
+  };
 }