+++ test/module/parse-stream.test.ts
@@ -1,18 +1,19 @@
 import { assert } from "chai";
-import { KipperCompiler, KipperStreams } from "../../src/";
 import { promises as fs } from "fs";
+import { KipperParseStream } from "../../src";
 
 const fileLocation: string = `${__dirname}/../kipper-files/main.kip`;
 
-describe("KipperCompiler.parse", () => {
-  it("KipperStreams.fromString", async () => {
-    let fileContent = (await fs.readFile(fileLocation, "utf8" as BufferEncoding)).toString();
-    let compiler = new KipperCompiler();
-    let stream = await KipperStreams.fromString(fileContent, "inline-stream");
+describe("KipperStreams", () => {
+  describe("KipperStreams.fromString()", () => {
+    it("Simple file initialisation", async () => {
+      let fileContent = (await fs.readFile(fileLocation, "utf8" as BufferEncoding)).toString();
+      let stream: KipperParseStream = new KipperParseStream("inline-stream", fileContent);
 
-    assert(stream.name === "inline-stream");
-    assert(stream.stringContent === fileContent);
-    assert(stream.charStream.sourceName === "inline-stream");
-    assert(stream.charStream.toString() === fileContent);
+      assert(stream.name === "inline-stream");
+      assert(stream.stringContent === fileContent);
+      assert(stream.charStream.sourceName === "inline-stream");
+      assert(stream.charStream.toString() === fileContent);
+    });
   });
 });